requirements.txt
psutil>=5.9
pygetwindow>=0.0.9; platform_system=="Windows"
pywin32>=306; platform_system=="Windows"
pynput>=1.7
watchdog>=4.0
Jinja2>=3.1
matplotlib>=3.8
pdfkit>=1.0
weasyprint>=62.0
schedule>=1.2
click>=8.1
platformdirs>=4.2
tenacity>=9.0
typing-extensions>=4.12
pytest>=8.3
pytest-cov>=5.0
reportlab>=4.2
Pillow>=10.3
pyautogui>=0.9
pandas>=2.2
seaborn>=0.13
cryptography>=43.0
customtkinter>=5.2


